add db layer for query command

+++ src/main/java/ru/akirakozov/sd/refactoring/servlet/AddProductServlet.java
@@ -18,7 +18,7 @@ public class AddProductServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
         String name = request.getParameter("name");
-        long price = Long.parseLong(request.getParameter("price"));
+        int price = Integer.parseInt(request.getParameter("price"));
 
         Database db = new Database();
         db.addProduct(name, price);
