+++ src/main/java/ru/akirakozov/sd/refactoring/utils/Item.java
@@ -2,9 +2,9 @@ package ru.akirakozov.sd.refactoring.utils;
 
 public class Item {
     String name;
-    long price;
+    int price;
 
-    public Item(String name, long price) {
+    public Item(String name, int price) {
         this.name = name;
         this.price = price;
     }
